Change ObjectComparison.cs
Added new field DayOfBirth of type DateTime to Person class for example of pros of realization using FluentAssertion and cons of realization using custom equality, where we need to do not forget to rewrite AreEqual method, because AreEqual useless when it does not report about our mistake ( mistake caused with day of birth of Vasili III in method TsarRegistry.GetCurrentTsar ). But also we need to edit all lines, where Person constructor is used, and it may be hardly. This may be fixed using Fluent Builder pattern.